requirements.txt
# Flask and extensions
Flask==2.0.1
Flask-Login==0.5.0
Flask-Session==0.5.0
Werkzeug==2.0.1

# AI and ML
google-generativeai==0.3.0
huggingface-hub==0.20.3

# Image processing
Pillow==10.0.0

# HTTP and API
requests==2.31.0
urllib3==1.26.15

# Server
gunicorn==20.1.0

# Security
itsdangerous==2.1.2

# Template engine
Jinja2==3.1.2
MarkupSafe==2.1.3

# Utilities
click==8.1.7
typing-extensions==4.7.1

# Session management
cachelib==0.9.0